+++ packages/interaction-kit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # interaction-kit
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [ac1256b]
+  - discord-verify@1.0.1
+  - discord-request@0.0.5
+  - discord-api@0.0.3
+
 ## 0.2.1
 
 ### Patch Changes